+++ examples/pathogen_evolution/pathogen_evolution.jl
@@ -23,8 +23,8 @@ function run(seed::Int64, t_vec::Vector{Float64})
         "pat_type",
         num_loci=4,
         possible_alleles="AB",
-        mean_effective_inoculum=1.0,
-        mean_mutations_per_replication=0.001,
+        # mean_effective_inoculum=1.0,
+        # mean_mutations_per_replication=0.001,
         contactCoefficient=s::String -> 1.0 + (0.1 * (4.0 - hamming(s, optimal_genome)) / 4.0),
         receiveContactCoefficient=s::String -> 0.0,
     )
@@ -36,6 +36,8 @@ function run(seed::Int64, t_vec::Vector{Float64})
         clearance_coefficient=1.0,
         contact_coefficient=1.05,
         receive_contact_coefficient=1.0,
+        mutations_upon_infection_coefficient=0.001,
+        inoculum_coefficient=1.0,
         pathogenFractions=jOpqua.pathogenFractionsProportionalFitness,
     )