+++ src/simulation/structs.jl
@@ -11,13 +11,13 @@ struct PathogenType
     num_loci::Int64
     possible_alleles::String
 
-    mean_effective_inoculum::Float64
-    mean_mutations_per_replication::Float64
-    mean_recombination_crossovers::Float64
+    # mean_effective_inoculum::Float64
+    # mean_mutations_per_replication::Float64
+    # mean_recombination_crossovers::Float64
 
-    inoculumCoefficient::FunctionWrapper{Float64,Tuple{String}} # takes seq argument, returns Float64
-    mutationCoefficient::FunctionWrapper{Float64,Tuple{String}} # takes seq argument, returns Float64
-    recombinationCoefficient::FunctionWrapper{Float64,Tuple{String}} # takes seq argument, returns Float64
+    # inoculumCoefficient::FunctionWrapper{Float64,Tuple{String}} # takes seq argument, returns Float64
+    # mutationCoefficient::FunctionWrapper{Float64,Tuple{String}} # takes seq argument, returns Float64
+    # recombinationCoefficient::FunctionWrapper{Float64,Tuple{String}} # takes seq argument, returns Float64
 
     verticalTransmissionCoefficient::FunctionWrapper{Float64,Tuple{String}} # takes seq argument, returns Float64
     # acquireResponseUponClearanceCoefficient::FunctionWrapper{Float64,Tuple{String}} # takes seq argument, returns Float64
@@ -29,8 +29,8 @@ end
 struct ResponseType
     id::String
     inherit_response::Float64
-    inoculumCoefficient::FunctionWrapper{Float64,Tuple{String,String,String,String}}
-    # takes host, imprinted, matured, and infecting sequences and returns Float64 coefficient
+    # inoculumCoefficient::FunctionWrapper{Float64,Tuple{String,String,String,String}}
+    # # takes host, imprinted, matured, and infecting sequences and returns Float64 coefficient
     verticalTransmissionCoefficient::FunctionWrapper{Float64,Tuple{String,String,String,String}}
     # takes host, imprinted, matured, and infecting sequences and returns Float64 coefficient
     infectionCoefficient::FunctionWrapper{Float64,Tuple{String,String,String,String}}
@@ -61,9 +61,9 @@ struct Pathogen
     # This is only useful for response lineage tracing, but not the simulation?
     sequence::String
     coefficients::SVector{NUM_COEFFICIENTS,Float64}
-    mean_effective_inoculum::Float64
-    mean_mutations_per_replication::Float64
-    mean_recombination_crossovers::Float64
+    # mean_effective_inoculum::Float64
+    # mean_mutations_per_replication::Float64
+    # mean_recombination_crossovers::Float64
     vertical_transmission_coefficient::Float64
     type::PathogenType
 end
@@ -123,9 +123,9 @@ struct PopulationType
     constant_contact_density::Bool
     constant_transition_density::Bool
 
-    inoculum_coefficient::Float64
-    mutation_coefficient::Float64
-    recombination_coefficient::Float64
+    # inoculum_coefficient::Float64
+    # mutation_coefficient::Float64
+    # recombination_coefficient::Float64
     vertical_transmission_coefficient::Float64
 
     host_num_loci::Int64